refactor
- перемещены файлы с функциями запросов клиента (агента)
- в хранилище метрик и counter добавлен RWMutex
