+++ internal/client/requests/sendMetric.go
@@ -1,4 +1,4 @@
-package client
+package requests
 
 import (
 	"fmt"
